# www_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the Laravel application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/languages": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Get list of Languages",
                "operationId": "1cb1ddff581b67412e8c4d53e8a5286d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "en"
                                            },
                                            "short_name": {
                                                "type": "string",
                                                "example": "Eng"
                                            },
                                            "full_name": {
                                                "type": "string",
                                                "example": "English"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Languages"
                ],
                "summary": "Store a new Language",
                "description": "Creates a new Language and stores it in the database.",
                "operationId": "storeLanguage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "short_name",
                                    "full_name"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Code of Language",
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "short_name": {
                                        "description": "Short name of Language",
                                        "type": "string",
                                        "example": "Eng"
                                    },
                                    "full_name": {
                                        "description": "Full name of Language",
                                        "type": "string",
                                        "example": "English"
                                    },
                                    "is_enabled": {
                                        "description": "Is enabled Language",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "order": {
                                        "description": "Order of Language",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Languages created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "Code of created Language",
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "short_name": {
                                            "description": "Short name of created Language",
                                            "type": "string",
                                            "example": "Eng"
                                        },
                                        "full_name": {
                                            "description": "Full name of created Language",
                                            "type": "string",
                                            "example": "English"
                                        },
                                        "is_enabled": {
                                            "description": "Is enabled Language",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "order": {
                                            "description": "Order of Language",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}